Adds support for custom thresholds
This lets you give approximate answers when it makes sense to do so.

For eg, the differnce between 6 years and "6 years, 2 days"
is so minor to be of no consequence.

(2 days is 0.09% of 6 years)

The threshold configuration lets you get approximate answers
by setting a minimum % (as a fraction). So if you set it to 0.01,
any values that are less than 1% of the previous term
will be silently dropped.

The default value is set to 0, so as to return accurate answers (and
remain backward compatible) by default